+++ Sources/UICollectionViewKit/Configuration/CategoryItemBackgroundConfiguration.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-public struct CategoryItemBackgroundConfiguration: Sendable {
+public struct CategoryItemBackgroundConfiguration: Sendable, Equatable {
     public var viewBackgroundColor: UIColor
     public var headerBackgroundColor: UIColor
     public var gridBackgroundColor: UIColor
@@ -17,6 +17,15 @@ public struct CategoryItemBackgroundConfiguration: Sendable {
 }
 
 extension CategoryItemBackgroundConfiguration {
+    public static func == (
+        lhs: CategoryItemBackgroundConfiguration,
+        rhs: CategoryItemBackgroundConfiguration
+    ) -> Bool {
+        lhs.viewBackgroundColor.isEqual(rhs.viewBackgroundColor)
+            && lhs.headerBackgroundColor.isEqual(rhs.headerBackgroundColor)
+            && lhs.gridBackgroundColor.isEqual(rhs.gridBackgroundColor)
+    }
+
     public static let `default` = CategoryItemBackgroundConfiguration(
         viewBackgroundColor: .systemBackground,
         headerBackgroundColor: .systemBackground,